# Fix panic in distributor due to early cleanup