--- a/cmd/mimir/help-all.txt.tmpl
+++ b/cmd/mimir/help-all.txt.tmpl
@@ -754,6 +754,8 @@ Usage of ./cmd/mimir/mimir:
     	[experimental] Enables the feature to forward certain metrics in remote_write requests, depending on defined rules.
   -distributor.forwarding.propagate-errors
     	[experimental] If disabled then forwarding requests are always considered to be successful, errors are ignored. (default true)
+  -distributor.forwarding.request-concurrency int
+    	[experimental] Maximum concurrency at which forwarding requests get performed. (default 10)
   -distributor.forwarding.request-timeout duration
     	[experimental] Timeout for requests to ingestion endpoints to which we forward metrics. (default 10s)
   -distributor.ha-tracker.cluster string