Fixes: https://github.com/grafana/mimir/issues/2266

Edit 2022-07-22:
Most of the code in `pkg/distributor/forwarding` has been re-written so looking at the diff between the old and the new code is probably not very helpful when reviewing, I recommend looking at the new implementation as if it were completely new.

In a conversation via DMs with @bboreham we concluded that it would be better to keep the forwarding logic as isolated from the distributor's `PushWithCleanup()` as possible, this should help to ensure that there are no race conditions or pool usage bugs.