@bboreham 2022-07-11
This feature seems incredibly complicated to implement, thus hard to see if there are any remaining bugs.

Conceptually it seems to me it might work better as a proxy in front of the distributor, either a separate process or a wrapper function in the code which strips out those metrics which are to be forwarded and passes the remainder to the function which sends to ingesters.