@bboreham 2022-07-19
Some good thoughts there; I wanted to pick up one point:

> The cleanup in the current PushWithCleanup could then be implemented as a middleware like:
...
> 	defer mimirpb.ReuseSlice(req.Timeseries)

The "cleanup", which is to reuse memory and avoid garbage-collection, has to wait until all calls from distributor to ingester have finished.  This is after the point where a `defer` would run, because we return after getting a majority (e.g. 2 out of 3) of responses, and some calls are still running.