@pstibrany 2022-07-20
This discussion is somewhat related to discussion we have in https://github.com/grafana/mimir/pull/2419.