@dimitarvdimitrov 2022-07-20
Pushed some example middleware on a different branch just to demonstrate what I meant 435daaf6b23971f6ce8915ea9266d13492b60a14 It might not compile, also not sure where to put the ctx yet