@replay 2022-07-22
This PR has too much history which goes back and forth between different approaches so the git history isn't useful for reviewing anyway, I'll squash & rebase.