@colega 2022-07-26
I see that `forwardSamples` is called only from `Push`, but the method that API uses is `PushWithCleanup`.

So this basically does not forward samples ingested through the API (`Push` is only used by the ruler as far as I can see).

Is that a bug?