@replay 2022-07-26
FYI force-pushing to rebase onto main because https://github.com/grafana/mimir/pull/2295#discussion_r930086446