@replay 2022-07-27
I think I addressed all the comments, @colega and @56quarters if you get a chance it would be great if you could take another 👀 🙏 .

The messages about "Helm <> Jsonnet Diff" seem to be because whatever job is posting them doesn't seem to support generics 🤦‍♂️