@replay 2022-07-28
I deployed this branch into our dev cell and I noticed a problem: It seems that some label values of forwarded time series get corrupted. I think this is due to our usage of "yoloStrings" in mimirpb, to fix this we'll have to copy strings by value :( working on it.