@replay 2022-07-29
@oleg and @56quarters  I pushed another commit which fixes the issue I described in my last comment, PTAL: https://github.com/grafana/mimir/pull/2295/commits/6ed53a31d9325a7dbdc6c06b2747e8f1564922ac

I don't like how complicated it is 😞 , it is pooling the byte slices which are backing the yolo strings created in `mimirpb`. If you have an idea how we can simplify this I would be very happy to hear it. Obviously, one way to simplify it would be to just use "normal" strings in `mimirpb.DeepCopyTimeseries()`, however this would increase the number of allocated objects by roughly about `7500%` according to the included benchmark, the current solution only increases the allocations by less than `10%`:

```
$ benchstat -delta-test none ./before after4
name                                                      old time/op    new time/op    delta
RemoteWriteForwarding/forwarding_and_ingesting-4            1.65ms ± 0%    1.74ms ± 0%   +5.13%
RemoteWriteForwarding/forwarding_and_no_ingesting-4         1.70ms ± 0%    1.81ms ± 0%   +6.13%
RemoteWriteForwarding/no_forwarding_and_only_ingesting-4    1.45ms ± 0%    1.38ms ± 0%   -4.65%

name                                                      old alloc/op   new alloc/op   delta
RemoteWriteForwarding/forwarding_and_ingesting-4            2.19kB ± 0%    2.52kB ± 0%  +15.15%
RemoteWriteForwarding/forwarding_and_no_ingesting-4         2.43kB ± 0%    2.60kB ± 0%   +7.08%
RemoteWriteForwarding/no_forwarding_and_only_ingesting-4      195B ± 0%      195B ± 0%    0.00%

name                                                      old allocs/op  new allocs/op  delta
RemoteWriteForwarding/forwarding_and_ingesting-4              27.0 ± 0%      29.0 ± 0%   +7.41%
RemoteWriteForwarding/forwarding_and_no_ingesting-4           28.0 ± 0%      29.0 ± 0%   +3.57%
RemoteWriteForwarding/no_forwarding_and_only_ingesting-4      5.00 ± 0%      5.00 ± 0%    0.00%
```