@replay 2022-07-29
@colega had a great idea how I can simplify the code by moving the yolo slice pool into `mimirpb`. This makes the code significantly easier to understand 🎉. I implemented it in the commit: https://github.com/grafana/mimir/pull/2295/commits/f10adf0290deebdde04664fe56f07b425aca81e6

The benchmarks are also looking good, comparing `9f74614` (when I noticed the yolo string issue) against `f10adf0`:

```
$ benchstat -delta-test none ./9f74614 ./f10adf0
name                                                      old time/op    new time/op    delta
RemoteWriteForwarding/forwarding_and_ingesting-4            1.76ms ± 0%    2.00ms ± 0%  +13.60%
RemoteWriteForwarding/forwarding_and_no_ingesting-4         1.87ms ± 0%    2.01ms ± 0%   +7.73%
RemoteWriteForwarding/no_forwarding_and_only_ingesting-4    1.48ms ± 0%    1.46ms ± 0%   -1.52%

name                                                      old alloc/op   new alloc/op   delta
RemoteWriteForwarding/forwarding_and_ingesting-4            2.29kB ± 0%    2.78kB ± 0%  +21.40%
RemoteWriteForwarding/forwarding_and_no_ingesting-4         2.45kB ± 0%    2.89kB ± 0%  +17.99%
RemoteWriteForwarding/no_forwarding_and_only_ingesting-4      199B ± 0%      192B ± 0%   -3.52%

name                                                      old allocs/op  new allocs/op  delta
RemoteWriteForwarding/forwarding_and_ingesting-4              27.0 ± 0%      29.0 ± 0%   +7.41%
RemoteWriteForwarding/forwarding_and_no_ingesting-4           28.0 ± 0%      29.0 ± 0%   +3.57%
RemoteWriteForwarding/no_forwarding_and_only_ingesting-4      5.00 ± 0%      5.00 ± 0%    0.00%
```